Move exception beans to their own subfolder
